backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image Processing
pillow==10.1.0
opencv-python==4.6.0.66  # Must be 4.6.0.66 for paddleocr compatibility
numpy==1.26.2
scikit-image==0.22.0
scipy==1.11.4

# OCR Engines
easyocr==1.7.0
paddlepaddle==2.6.2  # Updated from 2.6.0 for compatibility
paddleocr==2.7.3
pytesseract==0.3.10

# Math OCR
pix2text==0.2.3

# Gemini Vision API (fallback OCR)
# CRITICAL: Use recent version that properly handles list_models() generator
google-generativeai>=0.8.0

# Deep Learning (for TrOCR - optional, used in ocr_engine.py)
# CUDA-enabled PyTorch for RTX 4060 GPU
# Installed: torch 2.5.1+cu121, torchvision 0.20.1+cu121
# To reinstall GPU version: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# For CPU-only: pip install torch==2.1.1+cpu torchvision==0.16.1+cpu
transformers==4.35.2
torch>=2.5.1+cu121  # CUDA 12.1 version for RTX 4060 GPU
torchvision>=0.20.1+cu121  # CUDA 12.1 version for RTX 4060 GPU

# Document Generation
python-docx==1.1.0

# Utilities
matplotlib==3.8.2
pandas==2.1.3
python-dotenv==1.0.0